Parquet export IR
Exporting data through a CSV intermediary is subject to loss
of specificity and type info. This is particularly noticable
for read_pandas, where the resulting dataframe has every column
of type `object` and NULLs are indistinguishable from zero values.

I used a small hack to export data from Dolt into a DataFrame using parquet
instead of CSV. This requires the pyarrow dependency.

I left TODOs for improvements on the Dolt side that would make
this code cleaner and Dolt issues for the associated features.

There is one bug with NULL datetime values that I added a Dolt issue
for.